+++ src/tech/v3/dataset/metamorph_api.clj
@@ -26,7 +26,7 @@
 
 (def ^:private excludes-dataset
   '#{bind-> all-descriptive-stats-names major-version ->dataset ->>dataset column-map-m
-     mapseq-parser mapseq-rf})
+     mapseq-parser mapseq-rf dataset-parser})